fix: extract IconNotification to separate component
